chore: fix linting issues (#22)
## Description

Fix all error-causing linting issues. 

Required by #21
